Test now exercises function execution using id. Split Dunit test into multiple tests.
